Created wrapper header and initial wrapper interface
note code does not compile as of now due to dependencies not set up
correctly. The interface is intended to be a cpp header file that can be
included in a c file. Thus the interface will be able to make calls to
rapidJSON functions so if it is included in rapidJSON_wrapper.c it can
use rapidJSON's parse function. Definitely needs more work but it's
getting there.