Messing with linking all the files together
Note this still does not compile but I think it's closer to the final
product. There's not too much left to do besides figure out exactly how
to link the files and then implement the callback to the extern C parse
function from the to_value function in rapidJSON_wrapper.c. Almost
there!